requirements.txt
# Core scientific stack
numpy>=1.26
matplotlib>=3.8

# LLM providers (install at least one)
google-genai>=1.0        # Gemini provider
groq>=0.9                # Groq provider

# Output generation
python-pptx>=0.6.23      # PowerPoint presentation (Agent 8)
# python-docx is NOT required — Word reports use stdlib zipfile + raw OOXML

# DIgSILENT PowerFactory Python interface is installed with PowerFactory itself.
# Add the following path to PYTHONPATH (or sys.path) before importing powerfactory:
#   C:\Program Files\DIgSILENT\PowerFactory <version>\Python\<python_version>
